Move distributed stuff to a distributed file
